quant-idea-extractor/requirements.txt
# Core
anthropic>=0.40.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
aiosqlite>=0.20.0
loguru>=0.7.0
rich>=13.7.0
typer>=0.12.0
tenacity>=8.2.0
httpx>=0.27.0

# Collectors
praw>=7.7.1
arxiv>=2.1.3
feedparser>=6.0.11
beautifulsoup4>=4.12.3
trafilatura>=1.12.0
yt-dlp>=2024.1.0

# Dedup / ML
scikit-learn>=1.4.0
numpy>=1.26.0
scipy>=1.12.0
